# curl: set a 100K buffer size by default