--- a/tests/data/test1404
+++ b/tests/data/test1404
@@ -123,6 +123,7 @@ int main(int argc, char *argv[])
                CURLFORM_END);
 
   hnd = curl_easy_init();
+  curl_easy_setopt(hnd, CURLOPT_BUFFERSIZE, 102400L);
   curl_easy_setopt(hnd, CURLOPT_URL, "http://%HOSTIP:%HTTPPORT/we/want/1404");
   curl_easy_setopt(hnd, CURLOPT_HEADER, 1L);
   curl_easy_setopt(hnd, CURLOPT_HTTPPOST, post1);