@bagder 2017-04-24
Ah, forgot to update some tests. Will do.

Interestingly, changing `CURL_MAX_WRITE_SIZE` to 100K and a recompile generates even lower numbers, indicating that there's some glitches in the use of the larger buffer...