@jay 2017-04-24
I suspect this is unnecessary and you should back it up with some typical real world data scenarios. Even if there's a very slight difference in a corner case iirc the user can already set up to a 500k buffer size via rate limiting.

For comparison I ran your 80GB test in Windows. There is no discernible difference in a larger buffer size. I recall I discussed this with you or someone else before and didn't we reach the conclusion that this is only helpful for sftp? Also you may have blogged about it. Sorry my memory is foggy on this.

create file:
~~~
fsutil file createnew 80GB 0
fsutil sparse setflag 80GB
c:\gnuwin32\bin\dd if=nul of=80GB bs=1 count=0 seek=80GB oflag=append
~~~

bagder/curl-larger-buffer:
~~~
curl http://localhost/80GB -so NUL --write-out "time_total: %{time_total}"
time_total: 147.811000
curl http://localhost/80GB -so NUL --write-out "time_total: %{time_total}"
time_total: 145.939000
curl http://localhost/80GB -so NUL --write-out "time_total: %{time_total}"
time_total: 145.798000
~~~

parent you branched from, 7474418:
~~~
curl http://localhost/80GB -so NUL --write-out "time_total: %{time_total}"
time_total: 150.729000
curl http://localhost/80GB -so NUL --write-out "time_total: %{time_total}"
time_total: 149.683000
curl http://localhost/80GB -so NUL --write-out "time_total: %{time_total}"
time_total: 149.542000
~~~
~~~
curl 7.54.1-DEV (i386-pc-win32) libcurl/7.54.1-DEV OpenSSL/1.0.2k nghttp2/1.21.1
Protocols: dict file ftp ftps gopher http https imap imaps ldap pop3 pop3s rtsp smb smbs smtp smtps telnet tftp
Features: AsynchDNS Largefile NTLM SSL HTTP2 HTTPS-proxy
~~~
Server: Apache/2.4.18 (Win64)