@bagder 2017-04-25
I think this is an easy fix that makes some improvements at a very low cost. And it doesn't touch the library which makes it even less risky. To me, the numbers I've already shown is reason enough: curl will use less CPU and system resources this way. Sure, for regular network transfers it will be bottle-necked by the bandwidth but being lightweight is also a good virtue.

But there are some actual network speed reasons too, as you mention. SFTP in particular will get a serious boost with this.

## SFTP 4GB over localhost

old

real    0m32.391s
user    0m28.824s
sys     0m2.060s

new (larger buffer)

real    0m28.212s
user    0m26.728s
sys     0m1.332s

So, about 10% faster here.

## SFTP 4GB over a 200ms latency link

(test run by adding latency to 'lo')

old

Transfer rate: 310KB/sec

new

Transfer rate: 1930KB/sec

(The ratio in this speed increase is almost identical to the ratio of the buffer increase...)

## HTTP over a 200 ms latency link

As a comparison to see what it fares when we add latency to the mix and thus it isn't completely CPU bound anymore.

old

real    0m35.536s
user    0m0.144s
sys     0m0.176s

new

real    0m35.319s
user    0m0.072s
sys     0m0.124s

Virtually no speed difference, but slightly lower system load.