@jay 2017-04-25
sftp is the real winner here. I take issue though with it being less resources. You're trading one resource for another. sftp it's clearly worth it.